+++ src/ccSelectPolymerTypeHeadInCylinder.cpp
@@ -280,7 +280,7 @@ bool ccSelectPolymerTypeHeadInCylinder::Pack(const tguArgumentSequence& vArgumen
 {
 #if EnableCommandGroups == SimCommandEnabled
 
-    if(ccSelectPolymerTypeInRegion::Pack(vArguments) && vArguments.size() == GetArgumentTotal())
+    if(ccSelectPolymerTypeInRegion::Pack(vArguments) && static_cast<long>(vArguments.size()) == GetArgumentTotal())
     {
         vArguments.at(2)->GetValue(&m_X);
         vArguments.at(3)->GetValue(&m_Y);
