+++ src/ccSelectPolymerTypeHeadInPentagon.cpp
@@ -287,7 +287,7 @@ bool ccSelectPolymerTypeHeadInPentagon::Pack(const tguArgumentSequence& vArgumen
 {
 #if EnableCommandGroups == SimCommandEnabled
 
-    if(ccSelectPolymerTypeInRegion::Pack(vArguments) && vArguments.size() == GetArgumentTotal())
+    if(ccSelectPolymerTypeInRegion::Pack(vArguments) && static_cast<long>(vArguments.size()) == GetArgumentTotal())
     {
         vArguments.at(2)->GetValue(&m_XC);
         vArguments.at(3)->GetValue(&m_YC);
