+++ src/cgtAddCommandToGroup.cpp
@@ -293,7 +293,7 @@ bool cgtAddCommandToGroup::Pack(const tguArgumentSequence& vArguments)
 {
 #if EnableCommandGroups == SimCommandEnabled
 
-    if(vArguments.size() >= GetArgumentTotal())
+    if(static_cast<long>(vArguments.size()) >= GetArgumentTotal())
     {
         vArguments.at(0)->GetValue(&m_GroupName);
         vArguments.at(1)->GetValue(&m_CommandName);
@@ -306,7 +306,7 @@ bool cgtAddCommandToGroup::Pack(const tguArgumentSequence& vArguments)
         std::cout << "Command arguments for " << GetType() << " are: " << zEndl;
         std::cout << vArguments.at(0) << " " << vArguments.at(1) << zEndl;
 
-        if(vArguments.size() == argTotal+2)
+        if(static_cast<long>(vArguments.size()) == argTotal+2)
         {
             // Add their placeholder names of the command's arguments
             // to the local container. This works because all of the names
