+++ src/cgtExecuteCommandGroupOnce.cpp
@@ -194,7 +194,7 @@ bool cgtExecuteCommandGroupOnce::Pack(const tguArgumentSequence& vArguments)
 {
 #if EnableCommandGroups == SimCommandEnabled
 
-    if(vArguments.size() == GetArgumentTotal())
+    if(static_cast<long>(vArguments.size()) == GetArgumentTotal())
     {
         vArguments.at(0)->GetValue(&m_Name);
