+++ src/ctSetTargetInvisible.cpp
@@ -186,7 +186,7 @@ bool ctSetTargetInvisible::IsDataValid(const CInputData &riData) const
 
 bool ctSetTargetInvisible::Pack(const tguArgumentSequence& vArguments)
 {
-    if(vArguments.size() == GetArgumentTotal())
+    if(static_cast<long>(vArguments.size()) == GetArgumentTotal())
     {
         ctApplyCommand::Pack(vArguments);
         return true;
