+++ src/ctToggleTargetDisplay.cpp
@@ -178,7 +178,7 @@ bool ctToggleTargetDisplay::IsDataValid(const CInputData& riData) const
 
 bool ctToggleTargetDisplay::Pack(const tguArgumentSequence& vArguments)
 {
-    if(vArguments.size() == GetArgumentTotal())
+    if(static_cast<long>(vArguments.size()) == GetArgumentTotal())
     {
         return ctApplyCommand::Pack(vArguments);
     }
